+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/binding/DefaultBidirectionalConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.javafx;
+package com.canoo.dolphin.client.javafx.binding;
 
 /**
  * Created by hendrikebbers on 29.09.15.
