+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/impl/DefaultDolphinBinder.java
@@ -15,10 +15,10 @@
  */
 package com.canoo.dolphin.client.javafx.impl;
 
-import com.canoo.dolphin.client.javafx.BidirectionalConverter;
-import com.canoo.dolphin.client.javafx.Binding;
-import com.canoo.dolphin.client.javafx.Converter;
-import com.canoo.dolphin.client.javafx.DolphinBinder;
+import com.canoo.dolphin.client.javafx.binding.BidirectionalConverter;
+import com.canoo.dolphin.client.javafx.binding.Binding;
+import com.canoo.dolphin.client.javafx.binding.Converter;
+import com.canoo.dolphin.client.javafx.binding.DolphinBinder;
 import com.canoo.dolphin.event.Subscription;
 import com.canoo.dolphin.mapping.Property;
 import javafx.beans.value.ChangeListener;
