+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/impl/DefaultJavaFXBidirectionalBinder.java
@@ -15,9 +15,9 @@
  */
 package com.canoo.dolphin.client.javafx.impl;
 
-import com.canoo.dolphin.client.javafx.BidirectionalConverter;
-import com.canoo.dolphin.client.javafx.Binding;
-import com.canoo.dolphin.client.javafx.JavaFXBidirectionalBinder;
+import com.canoo.dolphin.client.javafx.binding.BidirectionalConverter;
+import com.canoo.dolphin.client.javafx.binding.Binding;
+import com.canoo.dolphin.client.javafx.binding.JavaFXBidirectionalBinder;
 import com.canoo.dolphin.mapping.Property;
 import javafx.beans.value.ChangeListener;
