+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/impl/numeric/LongDolphinBinder.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dolphin.client.javafx.impl.numeric;
 
-import com.canoo.dolphin.client.javafx.BidirectionalConverter;
+import com.canoo.dolphin.client.javafx.binding.BidirectionalConverter;
 import com.canoo.dolphin.mapping.Property;
 
 /**
