+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/stage/DolphinWindowUtils.java
@@ -1,6 +1,6 @@
 package com.canoo.dolphin.client.javafx.stage;
 
-import com.canoo.dolphin.client.javafx.AbstractViewBinder;
+import com.canoo.dolphin.client.javafx.view.AbstractViewBinder;
 import com.canoo.dolphin.event.Subscription;
 import com.canoo.dolphin.util.Assert;
 import javafx.event.EventHandler;
