+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/view/AbstractFXMLViewBinder.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.javafx;
+package com.canoo.dolphin.client.javafx.view;
 
 import com.canoo.dolphin.client.ClientContext;
 import com.canoo.dolphin.util.Assert;
