+++ platform/dolphin-platform-client-javafx/src/main/java/com/canoo/dolphin/client/javafx/view/FxmlLoadException.java
@@ -13,17 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.client.javafx;
+package com.canoo.dolphin.client.javafx.view;
 
-import javafx.beans.value.ObservableValue;
+public class FxmlLoadException extends RuntimeException {
 
-/**
- * Created by hendrikebbers on 28.09.15.
- */
-public interface NumericDolphinBinder<T extends Number> extends DolphinBinder<T> {
+    private static final long serialVersionUID = -586158083991150489L;
 
-    Binding toNumeric(final ObservableValue<Number> observableValue);
+    public FxmlLoadException() {
+    }
 
-    Binding bidirectionalToNumeric(final javafx.beans.property.Property<Number> property);
+    public FxmlLoadException(String message) {
+        super(message);
+    }
 
+    public FxmlLoadException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public FxmlLoadException(Throwable cause) {
+        super(cause);
+    }
 }
