+++ platform/dolphin-platform-client-javafx/src/test/java/com/canoo/dolphin/client/javafx/impl/BidirectionalListBinderTest.java
@@ -15,7 +15,7 @@
  */
 package com.canoo.dolphin.client.javafx.impl;
 
-import com.canoo.dolphin.client.javafx.FXBinder;
+import com.canoo.dolphin.client.javafx.binding.FXBinder;
 import com.canoo.dolphin.collections.ObservableList;
 import com.canoo.dolphin.impl.collections.ObservableArrayList;
 import javafx.collections.FXCollections;
